useSearchParams vercel fix 2
separated and simplified page.tsx and moved the functionalities to notes-content.tsx
